(#478) First prototype of OnDemandContentLoaderManager
- Implement OnDemandContentLoaderManager + content loader prototype.
- Remove all onViewAttachedToWindow/onViewDetachedFromWindow in PostCell (and other Post classes) and PostAdapter. The reasoning is it's bad to use those two callbacks with RecyclerView because they
have separate lifecycles. So when using onViewAttachedToWindow/onViewDetachedFromWindow you may end up in a situation when onViewDetachedFromWindow was called, you nulled out all the callbacks but
for RecyclerView that we might still be alive so you will end up with either the NPE or empty view.
- Replace ListView with RecyclerView in PostRepliesController because ListView would sometimes bind the same view twice which would break content loading subscription/unsubscription. And there was no
way to manually recycle replies when closing that controller.