(#478) Database optimizations + some fixes for 2ch.hk
Change the DB, add ChanThread column (which depends on ChanBoard column).
Possible fix for "Thread has no op" error since "op" parameter is not reliable for determining whether a post is OP or not (we now use "parent" parameter which is always 0 for OP and opId for every other post in a thread).