(#332) Fixes for the posts table
Fix a bug where OP could be saved with opId == 0 (which would potentially break something in the future).
Fix a bug where it was impossible to save two posts that belong to different threads (or even sites) if they have the same id (because postId was the primary key, now it's postId + threadId).