Implement WIP version of comment parser for FoolFuuka + bugfixes
FoolFuukaCommentParser can now parse quotes/links/replies etc (but it's still not perfect).
Mark posts as delete and images with "From archive" label if an image/post is from archive.
Fix a bug where both the original and archived images would be shown in the catalog's PostCells.
Add a ChanPostReplyEntity table to store posts' replies