Bug fixes
Fix a bug in archives media parsing exception. Was using a wrong link in json - "remote_media_link" instead of "media_link". The problem is "remote_media_link" can be null even though the image actually exists and can be opened. And sometimes it stays null for who knows how much time. "media_link" on the other hand is never null is works as expected.