Implement PostsCache with fixes for stale catalog info
The problem was in how we store posts and original posts in the HashMap. We were using different keys for that - CatalogDescriptor and ThreadDescriptor. But it turned out that when updating the original posts we would use wrong keys so that we never update the actual data. Now posts and original posts are stored in two different HashMaps and when we want to get an original post we get the both and the merge into one (We take some info from original post and set it into the normal post). We do that because catalog and thread APIs may contain different data for stuff like lastModified, repliesCount etc. for posts and original posts.