Implement smart archive selection.
So how does it work. After every archive fetch (when we fetch posts from an archive) we store the fetch result in the database (we store a flag whether the fetch was successful or did it end up with some kind of an error, and if there was an error we also store it's message for debugging purposes). Then when we want to fetch posts from an archive, we search for every archive that supports a board we are currently on, the we filter out all archives that do not store files for this board. Then we check the previous 5 fetch results. If all fetches ended up with an error for an archive - we skip it and. If all fetches ended up with an error we return null, which means we won't fetch anything. If there are couple of archives with some successful previous fetches, we sort them by the amount of successful fetches and then again select the most suitable one. The use can override the behavior when all archives have no successful fetches by clicking the "reload" button in the toolbar or by pressing the bottom thread status view. Both of them will ignore a situation when all archives are "dead".