Do not use posts list we get from the server to sort posts when loading posts in a thread.
Because that list won't contain deleted posts and we will never load
them from the DB in such case. Instead, we should load whatever there is
in the database for the thread and then sort it by ourselves. But when
loading catalog posts use the server post list for ordering like we did
before.