Fix couple of bugs related to parsing, merging and storing of archived posts:
- Fix a bug in PostUtils.postsDifferFast (now it's called
shouldRetainPostFromArchive) where we would retain an archived post if
it's imageUrl (or thumbnailUrl) is not the same as one in cached post.
This is incorrect because it will overwrite a good post with it's
archived version when we don't want that. The correct logic here is to
use the one post that has both imageUrl and thumbnailUrl not set to null.
If both posts have them then use the regular post (not the archived one).

- Fix a bug in ChanLoaderRequestExecutor.merge() function where we never
add posts from the archives to the result list which ends up in them
never being cached in the database.