Many improvements for posts that are stored in the database.
- Now we store posts for every archive type as well as the normal posts.
But, we don't actually store all archive the posts. We only store those
archive posts that are more valuable than their regular version (one of
the condition to be a more valuable post is to have more images than
the regular post. This handles the case to use archive post when the
regular post had it's image(s) deleted from the server).
- Use archiveId in posts and post images instead of "isArchived" flag.
- Introduce PostIdEntity - a post id that is composed from many other ids
(threadId, archiveId, postNo and postSubNo). This is a universal key
to find a post in the database.
- Introduce ArchivePostDescriptor (which is not used anywhere for now)
that has a field "postSubNo". It will be used for archive's ghost posts.
(By default it's 0 for all non-ghost posts).
- Disable post cache (for now).
- Add couple of tests to cover the most complex cases for
ChanPostLocalSource