Fix a bug where we would not reload deleted posts from the database that we fetched from an archive when ArchiveManager.getBestPossibleArchiveOrNull() returns null.
This is because when reloading posts from the disk we need 2 archive ids.
The first one is mostly always 0 and the other one should be an id of
the archive where we had fetched the posts from. But after fetching the
posts once getBestPossibleArchiveOrNull() would start returning null
afterwards (until the cooldown dies down) so because the archiveId is
null we would never reload that archive's posts. The solution is to store
the last archive descriptor returned by getBestPossibleArchiveOrNull()
until the next time it recalculates the scores and use it. Each thread
will have it's own archive descriptor.