Fix couple of bugs.
- Fix a bug where navigation history would be in "loading"
state until we open any thread/board. That looks strange
when we open the app the very first time when we have no
configured boards. The spinner will be rotating until we
add the first board. Now it's fixed and it will show the
default message instead of the spinner.
- Fix a crash when exporting setting. Was cause due to an
attempt to pop navigation controller.