Fix a really weird bug that occurs after setting SYSTEM_UI_FLAG_LAYOUT_STABLE flag.
For some reason after settings that flag, and then removing it, onLayout() methods
stop getting called. For now, the solution is to not use that flag whatsoever. This
is a hacky solution but it works with minimal amount of drawbacks (like the only
one I noticed is the Toolbar jumps to the top of the screen when opening an image).

Fix some other bugs:
- Disable drawer's longtap-top-show gesture when Bookmarks or Settings controller
is opened.
- Fix a bug where the Toolbar won't be updated when opening a thread via Bookmarks.
(it will stay in "catalog" mode).