Cloudflare image preloading for 4chan.
Add an option to force Cloudflare to preload all images (excluding inlined images) in a thread. No
images are downloaded locally, only HEAD requests are sent. It turns out that sending a HEAD
request to an image will force Cloudflare to preload it which makes the actual image download way
faster because it will be served from the Cloudflare servers and not from 4chans'.