Add new span types: ColorizableBackgroundColorSpan and ColorizableForegroundColorSpan.
- The difference is they don't contain a theme color itself, instead they contain a color id which
then can be used to get the color itself from the current theme. This makes it easy to update
colors of spannable strings because before the change, since the spans are stored in the DB we
had to reparse posts to apply new theme color which is ridiculously inconvenient and complex. Now
we don't need to do that since all span colors are resolved by their ids. Now theme change is fully
dynamic!
- Introduce "normalControlColor" and "controlDisabledAlpha" and use them instead of lightening and
darkening colors.