Fix a really shitty bug that's been around since early Kuroba days.
Whenever you opened a thread and scrolled to bottom right away, sometimes, it would scroll you back
to top. Scroll back to bottom and clicking the PostStatusCell would scroll you back to top again
and it would continue until you switch to another thread or restart the app. Apparently, it was
happening because RecyclerView tries to auto scroll back to a focused view. This is absolutely
ridiculous and not obvious. Hopefully, this shit is finally completely fixed.