Fix a crash + some other fixes
- Fix a crash when trying to bookmark a thread that belongs to a site that was disabled before
updating to a version that runs a migration that creates default bookmark groups. In this case
default groups for disabled sites won't be created so once there is an attempt to create a new
bookmark it will fail with FOREIGN KEY error. Now we always pre-create default groups upon app
start.
- Fix some colors for EpoxyExpandableGroupViews.
- Remove some redundant layouts (tablet versions of bookmark views) because we can just set
tablet parameter in runtime.
- Add some comments.