Partial thread loading.
It's now possible to load a part of a thread if needed. It's useful when you want to show to the user
a preview of a thread (Original post + 5 last posts of a thread) and don't want to pollute the cache
and the database with the rest of thread's posts (In other words you only want to cache in memory
and database only those 6 posts, not the whole thread).

We will use this feature when opening external threads or threads in archives.