Fix a bug where post content won't get updated with extra info for OP when opening a thread.
Basically, this was happening because we would consider original post that we see in a catalog
and original post that we see in a thread as the same post so when we open a thread and bind
this original post (that we had just bound while in catalog) we won't redraw it.

Also popup posts in reply chain can now be updated with new info too.