Fix a bug where it was possible that the thread cache could contain two original posts.
That happened when we loaded original posts from the database and then after a fresh server
update we would add those new posts into the cache. The problem was that when loading original
posts from the database we would use CatalogDescriptor for PostDescriptor and in all the other
places we are using ThreadDescriptors even for catalog posts. Those are not the same thing so
all "equals()" methods would return false and that's why we would store both of them.

Also, reduce debouncer time intervals in BookmarkWatcherCoordinator so there is enough time to
enqueue background work.