Fix a bug where when opening a thread via external link we would not scroll to a marked post (if the link contained it).
That happened because of race condition between a normal thread loading and quick thread loading
caused by themeEngine calling refreshViews() which called quickReload() in ThreadLayout.
So we loaded the same thread twice: once with the information from the database and the second
time with the fresh thread info from the server. And since when we were loading the thread
with cached posts we could some times not have the marked post cached (because of periodical post
deletion from the database upon app start).