Fix 4chan.org posts not appearing in threads when sending post with images.
The reason was a very strange 4chan.org behavior when using any User-Agent that is not completely
made up (The ones you can find by googling "firefox/chrome latest user agent". For some reason
4chan won't show posts made with those user agents even though the server sends back a successful
response. This is either a 4chan bug or it was done intentionally. The fix is to use a completely
made up user agent.