Fix couple of bugs.
- Fix thread's OP not getting marked as own post after creating a new thread.
- Fix not being able to create a thread on 2ch.hk
- Fix a really nasty bug where right after creating a new thread, when auto-bookmarking that thread
the method that handles success posting would be interrupted right in the middle and then canceled.
That was caused by calling unbindDescriptor() which calls job.cancelChildren() which caused this to
happen.
- Fix a bug where creating a thread with "Bookmark on post" option turned on would not auto bookmark
the thread. That happened because we were not pre-creating an empty the in the database which resulted
in us having no databaseThreadId which resulted in bookmark grouping throwing an NPE thus canceling the
auto-bookmarking.