Further work on filter watches.
- Fix a bug that was introduced when bookmark groups were introduced. A wrong databaseId was used
for ThreadBookmarkGroupEntryEntity's ownerBookmarkId (ownerThreadId was actually used instead of
threadBookmarkId).
- Get rid of reloadBookmarkFlag in EpoxyGridThreadBookmarkViewHolder/EpoxyListThreadBookmarkViewHolder.
Having this thing is very bad since by using it all the epoxy models will always be rebuilt even when
it's not needed. That flag was used as a workaround for a strange behavior where a click listener
lambda that was set on epoxy view would have a wrong captured value (if it were capturing any). The
bug happened because we didn't use "EpoxyAttribute.Option.IgnoreRequireHashCode" which should tell
epoxy to always rebind the lambda with fresh captured values.