Fix new thumbnail caching mechanism when highResCells setting is turned on.
The problem was that with highResCells setting turned on we would use the original image url instead of
thumbnail url, load it and then cache the result. Since both images that are being loaded via ImageLoader
as well as full resolution images use CacheHandler that image would be cache so when we try to load full res
version we would load cached thumbnail instead.