ReplyLayoutGestureListener fixes.
When a swipe is detected we need to check whether the view under
the touch position (start of swipe) can be scrolled towards the direction of swipe (up/down). If yes,
then we need to ignore such gesture. Otherwise this will lead to nasty bugs like swipe up/down being
triggered while we scroll any scrollable view (in case of ReplyLayout it may be RecyclerView or
NestedScrollView or even EditText).