More work on ImageLoaderV2.
- Fix a bug where two or more image load requests with the same url could be executed at the same
time which could corrupt the image on the disk (because they all could start writing at the same
time). Now if another request is detected while there is one active, the other one will be canceled
but it's callback will be added into a map, and in the end all callbacks will be notified once we have
the result.
- Force AlbumViewCell's ratio to be in the range 0.4f..2f