Fixes for laying out post with multiple images.
The problems was that we needed to know the full post cell width before it could be available
(before adding any thumbnails to the post view). The solution is to use doOnPreDraw + cache the result
width since it should be the same for the rest of the posts. But we need to separate post width for
1. Catalog posts, 2. Thread posts, 3. Popup posts since all of them will most likely have different
post cell widths.