Use a separate instance of FileManager for ImageSaver.
The current problem is that FileManager when in SAF mode is slow as hell. It also has all public methods
synchronized on the instance of FileManager. When downloading batches of images the FileManager will
be locked most of the time. This will make the rest of the app slow as hell since it uses the same
FileManager. The solution is to use a separate instance of FileManager in ImageSaver so it doesn't
block the other FileManager and thus doesn't hang the whole app.