Bugfixes.
- Due to a weird behavior of InstalldNativeService (sometimes it may start clearing the
Context.cacheDirectory while the app is running) we need to be able to recreate directories
that are located in the cache directory whenever it's cleared by the InstalldNativeService.
For now it's implemented for `file_chunks_cache` directory.
- Do not crash when an empty list is passed into BookmarksManager.deleteBookmarks().
- Fix "Standalone coroutine canceled" error when quickly swiping through the media viewer's view pager.
- Fix a bug where video playback is not restored into the "playing" state when quickly swiping through
videos in the media viewer.
- Fix couple of bugs that appear on slow network connections when opening media viewer (1. MediaViewer is
not shown until the initial media's preview is loaded. 2. Can't start swiping through viewpager until
the initial media's thumbnail is loaded).