Migrate crashlogs, anrs, file_chunks_cache, filecache directories from the Android cache directory to files directory.
The reason for this migration is that sometimes Android may start cleaning up the app's cache directory
even when the app is running (and even if the device has enough of disk space) which breaks everything
related to CacheHandler until app restart. We don't want this to happen so from now on all the cached
files will be located in the "files" directory. To clear app's disk caches (media and exoplayer)
there is a button in the developer settings (application's developer settings not Android's).

This weird Android behavior was one of the reasons of the bug where images sometimes would stop loading
until app restart.