Some improvements for post comment shift option when used in combination with post thumbnails being on the right side.
Basically, instead of checking whether all the first N lines of a post comment fit between the post
left side and thumbnail left side and then deciding to shift the comment up or not we now check how
many lines fit sum the fitting lines' offsets and then use it as a top margin with barrier excluding
the thumbnail. In other words the combination of post comment shift and thumbnails on the right side
is now the best possible combination and makes the most compact posts possible.
Switch the default 2ch.hk captcha from V2NOJS to DVACH_CAPTCHA.