Fix marking one of more bookmarks as fully read (read and close all notifications) not being persisted which would restore the read post counters to being fully unread on the next bookmarks update iteration.
The problem was in threads that had 'lastViewedPostNo' equal to 0, meaning we never visited that thread.
When marking bookmarks of such threads we use 'lastViewedPostNo' to figure out how many unread posts
we have on the next update iteration and since it's 0 we consider all thread posts unread. Now the
'threadLastPostNo' is introduced so whenever we mark a thread as fully read we simply set 'lastViewedPostNo'
to be 'threadLastPostNo' which fixes the issue.