Performance improvements when adding/updating thread posts.
- In the previous version we would always resort all thread posts if there was at least one added,
deleted or updated post. We would also recalculate the whole post reply chain on each thread update.
This could potentially take a lot of time in threads with lots of posts 1k+ and more. Now we only sort
the post if we detect that they are unsorted (which should never happen normally, unless the
image board is bugged by itself). We also calculate post reply chain incrementally basically
only for the new posts, not all of them like before. This should reduce the time we lock the
ChanThreads when updating them which should reduce the ANR reports.
- ChanPost repliesTo/repliesFrom fields are now PostDescriptors instead of simple longs which should add better
support for ghost posts in the future.