Rework how post name/posterId/tripcode/mod capcode is formatted.
Earlier, all that stuff would be formatted during the parsing stage, concatenated together and then
assigned to ChanPostBuilder.tripcode which is absolutely incorrect and makes everything unnecessary harder.
Now all that stuff is stored where it's supposed to be and span assignment is done in PostCellData instead of
during post comment parsing.