Split media disk cache into multiple disk caches.
The idea here is to store media displayed on different screen separately so that it's not evicted
from the cache because it's filled with other media. For example, let's take the bookmarked threads'
thumbnails. Before this change they were stored together with the other media, like post full images.
So it would be evicted from the cache just because you viewed lots of images. Now lets say the
bookmarked thread got deleted from the server. Now the bookmark thumbnails shows "Not found" which
is not desired. With the new cache implementation bookmark thumbnail won't be evicted from the cached
because it's filled with too many post images. It will only be evicted if it's filled with too many
other bookmark thumbnails. So the cached thumbnail will stay cached for much longer so you will be
seeing "Not found" way less frequently. There are 7 types of media caches now, each existing in it's
own separate directory:
- ThreadDownloaderThumbnail (local archive screen thread thumbnails).
- BookmarkThumbnail (bookmarks screen thread thumbnails).
- NavHistoryThumbnail (navigation history screen thread thumbnails).
- SiteIcon (all sites icons, like favicons or site flags).
- PostMediaThumbnail (post thumbnails).
- PostMediaFull (full size post media).
- Other (stuff like apk updates, media downloaded when importing it from external url, searx image cache, etc.)

They all take their own percent of the total allowed disk cache size. Most of them take 5% of the
disk cache size, exceptions are: PostMediaFull which takes 65% and Other which takes 10%. The percent values
will probably be adjusted in the future. For now you can't change them by yourself they are hardcoded. The
ability to change them may get added in the future.

This should also increase the disk cache access speed since files are now located in separate
directories the directory query should take less time due to having to query less files at a time.

You can clear all the disk caches separately now on the Developer settings screen.