Add ability to open drawer by dragging from left to right on the thread screen.
A new setting was added: viewThreadControllerSwipeable. Enabled by default it controls whether the
thread screen can be swiped away. If it's disabled then instead of the thread screen being swipeable
you will be able to open the drawer by performing the drag from left to right gesture from any
point of the screen (similar to how you can open the drawer from the catalog). If it's enabled
then you can open the drawer the same was as before - by longtapping the left screen edge and pulling
it to the right. viewThreadControllerSwipeable setting doesn't work in SPLIT layout mode.