Fix cloudflare clearance cookie not being remembered when using 2ch.life domain.
The are multiple reasons:
1. A lot more 2ch.hk endpoints have cloudflare checks.
2. Tons of requests being sent (and then fail) most of which are from the PagerRequestManager.

This creates a race condition which may end up in a situation where we after receiving the cloudflare
cookie we start sending many requests some of them after the cookie was received some of them before
so as soon as any of the requests (that has no cookie) fails we remove the cookie from the settings.
And this is what causing the endless cloudflare pop ups.