Change the logic that decides whether full media/thread media should be downloaded or not.
When imageAutoLoadNetwork or videoAutoLoadNetwork (also thread media download) setting is set to ChanSettings.NetworkContentAutoLoadMode.UNMETERED (ChanSettings.NetworkContentAutoLoadMode.WIFI previously) we now check if the current network is unmetered or not instead of checking if we are connected to WiFi because some WiFis can still be metered. If it's metered then we don't download full media.