Make 4chan post acknowledgement check more durable.
Try to look for the post 5 times with increasing timeouts (attemptIndex * 1 second). In addition to that, check the current catalog for any post with id greater or equal to the post that the server had returned to us. If it exists, and it is not in the same thread that we expect it to be, then fail posting early because it's definitely not our post.

The 2 downsides here are:
 1. On slow boards the amount of time it will take to make a post will increase by ~15 seconds + the time it takes to make a request.
 2. When two people post in the same thread at the same time, the server eats your post and the other post has post id that we expect then there is a high chance that we will assume the other post is ours because there is no simple way to check if it's actually ours or not. Comparing by comments won't work that easily (because of markdown) and the only other idea I have is to show that post to the user to check if it's actually his post.

 Maybe I will try to implement the latter idea if there is still too many false positives with the current solution.