fix: Lots of captcha fixes.
First of all, use the same function to extract domain/host from url everywhere when dealing with cookies. Difference in those will result in incorrect cookies being used for captcha.
Next, process the other two (new) CloudFlare cookies: "_tcs" and "__cf_bm". I guess they were added not that long ago.
Additionally, use a proper way to construct a single "Cookie" header from multiple subcookies. Introduce CookieBuilder for that. Ensure that we don't have multiple instance of the same cookie in the header. Replace the old cookie with new value when the same cookie is added into the builder.
Lastly, "Only remove expired WebView cookies" setting is now enabled by default as it seems to help (at least when I tested this locally).

After testing this locally, it seems like this completely removes the captcha timer after the IP changes, except for the very first time you request the captcha (not sure if it's normal or not).

Feedback is encouraged!