fix: rollback onlyRemoveExpiredWebviewCookies default value to false.
The reason for this is it sometimes makes it impossible to load the captcha. CloudFlare challenge is getting solved right away, captcha loads but there is another CloudFlare challenge. Not sure what causes this. This setting might work for some people but not for others. Feedback is needed to figure out whether to keep it or remove it.