fix: set onlyRemoveExpiredWebviewCookies's default value back to true.
Okay, I think I managed to solve the problem with not being able to load the captcha after passing the CloudFlare check. BUT! It still gives 60 seconds timeout after you change the IP. I give up, I don't know how to fix this. The website sends a lot more data, but I think this data is generated on 4chan client side and I don't really want to dig through that. I will leave it like this for now.