feat(captcha): captcha notifications.
When requesting captcha and it returns a rate-limit error, the captcha window can now be closed. When the captcha becomes available, a notification will be shown. Notification won't be shown if the captcha window is opened (it will be auto-reloaded).

Note: this only works if the app is active/minimized. If you actually close it it will stop tracking the captcha.